requirements.txt
streamlit-lottie
streamlit
pillow
plotly
pandas==1.5.3